ec2: add AWS EC2 support and module docs
This adds the ability to launch instances on AWS EC2. This
includes the ability to create custom VPCs, add ENI and EBS
devices to instances, and take snapshots of instances to
create images from them.

In addition, documentation for the modules was added all in
the Google docstring format. Sphinx support was added to
generate the docstrings as well.