docs: Create custom EC2 documentation
Creates an EC2 documentation page with steps on how to use the API and
configuration required for functionality.

Breaks the docs out of the main Makefile and places them in their own.
It also cleans up the main documentation page, adds the contributing
and API pages. As well as adds Markdown support to docs.