lxd: Add LXD support, docs, and examples
This adds the ability to launch instances on LXD. This includes the
ability to specify an image remote, storage, network, type, etc. The
ability to also create snapshots of the instances and then launch
instances from those snapshots or roll back to them.

An example code page as well as full API documentation for the LXD
cloud are also included.