Google Compute Platform: being support
Putting on hold as there are currently two different APIs available. The
first is the older, now in maitenance mode REST API that comes from
google-api-python-client.

The newer one that is still under development is google-cloud-python. I
would prefer to use the latter one, but untill it supports instance
launching from the Compute Platform it is not useable.