ec2: fix reboot
The reboot command initiates reboot on the OS itself not via the
EC2 infrastructure. Therefore, I cannot wait for instance to reach
a different state. As a result, we wait for a hardcoded period of
time for the system to come down and then start the standard wait
code to wait for cloud-init to finish and the system to be
available.