defaults: SSH Key and Image Release
Instead of requiring the user to always specify the SSH key, it
assumes his or her username + key found in the typical location.

The image release will also be set to default to the latest Ubuntu LTS
release.

These are sane defaults that should speed up development and enhance the
end user experience.