exceptions: removes custom exceptions from codebase
The exceptions code was poorly thoughtout and uncessaryily complex. It
is far better for now that we allow the libraries we use to throw their
own exceptions and allow the user to deal with those or check the return
codes of commands themselves.

This removes all the custom exception code and expects the user to do
the right thing.