Initial Google Compute Engine Support
This adds the inital support for GCE to pycloudlib. This enables
specifically the authentication mechanism with GCE, the ability to
search for daily Ubuntu images, and finally launch an instance. There
are no additionally support functions.

This does re-write how SSH keys are handled by clouds in order to make
it a bit more generic and allow for the GCE mechanism of declaring the
SSH key on boot, rather than needing to upload a key beforehand.