EC2 instance.restart(): consider a down instance as rebooting
Commands sent to an EC2 instance for which a reboot request has been
submitted may fail in unexcepted ways. Here we catch SSHException, which
is thrown by execute() on failure, and consider the instance as down
because a reboot is in progress, which is what we want. The while loop
can be exited and the rest of the wait logic is delegated to wait(),
which will in turn call boto3's methods.