Add support for Azure pro images
Azure pro images requires additional parameters to be provisioned.
We require the offer, publisher and sku of an image to compose a
'plan' dict with those fields. For snapshot images, we do not
have that information available in the image id. To better handle that,
we have updated the registered_images dict to hold such information.
This allows us to launch pro images from snapshots images as well.

Also, we have updated how we parse the image_id for information and
added the region as parameter to Azure cloud creation.

Fixes: #14