instance: allow KVM/LXD commands to exit non-zero
`pycloudlib.util.subp`'s default behaviour is to raise an exception if
the command being executed exits non-zero.  Before this commit, this
meant that executing a command against a LXD or KVM instance that exited
non-zero would raise an exception (instead of returning a `Result`
object as other instances do).

Fixes: #30