Fix regex that extracts release from image id (#91)
On LXD, we use a regex to extract the ubuntu release
from a image id. We try to find either the base release
name on the image id, like focal, or the version number,
like 20.4. However, if the image id has a timestamp on it
we could accidently match the timestamp value instead
of a valid version string. The reason for that is because
we are not scaping the dot part of the version string