Set service account email for gce instance
When we are launching an image, we do not bind it
to a service account, even if we are using a service
account credential to launch the images. This is problematic
because some of our vms need that to work properly. For example,
the recent gcp pro machines requires that bind to allow
uaclient to auto-attach on them. To fix that, we are now
performing that bind if the user provides a service_account_email
to the gce client