lxd/cloud.py: move image defaulting up call stack
If a remote is not specified for a LXD image, we default to the
ubuntu-daily remote.  Before this commit, this happened in
_BaseLXD._prepare_command, which meant that it had not happened when
LXDVirtualMachine._prepare_command was called.

This commit moves it up the call stack to happen in _BaseLXD.init, which
is (a) implemented only on the super-class, and (b) must be called for
any public API users to create a LXD instance.