lxd/cloud.py: use `lxc image info` in release determination
Both images: and ubuntu[-daily]: Ubuntu LXD images are published with
metadata that describes the OS and release of the contained image.  This
metadata is retained across image snapshots.

Instead of relying on naming conventions and regexes to determine the
release of a LXD image, this commit calls `lxc image info` and parses
the emitted YAML to find the apropos release.

Fixes: #95