lxd/instance: use SSH by default, allow switching to `exec`
Analysis performed (and captured in #112) indicates that accessing LXD
instances via SSH rather than via `lxc exec` leads to faster test
execution, and also brings LXD's default behaviour in line with the
default behaviour of other clouds.

This commit introduces a `execute_via_ssh` parameter and instance
variable, which are used by `execute()` to determine whether to connect
to the instance via SSH, or use `lxc exec`.  The default value of
`execute_via_ssh` is `True`.

(Prior to this commit, we would execute via SSH but only if we had a key
manually set against the instance, leading to inconsistencies in
behaviour depending on local configuration.  This also addresses that
inconsistency, and removes code that was required to support it.)

Fixes: #112