Drop support for xenial vstock image
Currently, LXD xenial virtual machines use a special image
to allow supporting installing lxd-agent on the vm and
run commands with lxc exec. Since we are now running all the commands
with ssh by default, we don't need to support that xenial image anymore.
The implication now is that xenial vm will not be able to run lxc exec
commands