Update naming convention for LXD vm profiles
Currently, when we update a LXD vm profile we may have not see it
reflected on subsequent runs because we do not generate a new profile
if there is one with the same name on the system. To avoid that, we
are now adding a version suffix to the profile name while also creating
a md5sum test that will fail if we update the profile content but do not
update the md5sum in the test.

Fixes: #65