Improve checking if LXD profile already exists
Currently, when we try to verify if a profile list
already exists, we run the lxc profile list command
and search if a given string name is in the output.
This approach is problematic because if the profile
name we are searching is a substring of another profile
name, we will say that the profile already exists, which is
not true. We are now updating that logic to fix
that