Add banner timeout for paramiko connect
We are observing that a lot of tests that use pycloudlib are
failing to due errors reading SSH protocol banner. Since we
cannot consistently reproduce those errors, we believe this should
be related to a timeout issue. To address that we are now setting
the banner_timeout variable when we run paramiko connect method