lxd/instance: raise NotImplementedError if console log unsupported
LXD VMs do not support fetching the console log.  This commit modifies
`LXDInstance.console_log` to detect the error LXD emits and `raise
NotImplementedError` in its place, to match the behaviour of other
clouds lacking this support.

(This method is preferable over detecting when we're using a LXD VM and
pre-emptively raising `NotImplementedError`: if LXD adds serial console
support to VMs, pycloudlib will be able to use that functionality
without code changes.)

Fixes: #130