ec2: wait for console log output
On EC2, console log output is not available immediately after launching
an instance: we've seen it take up to 2 minutes for the API to start
returning content.  Instead of, arguably incorrectly, claiming that
there is no console output in these cases, we retry for up to 3 minutes
until we get _some_ content.