Move all SSH retry logic to "_wait_for_execute()"
Before this change we would continously retry for 10 minutes wrapped
in a call to retry 10 times. 10*10=100 minutes for a single connection.
The only time this constant retrying should be necessary is if a cloud
instance is booting. The connection timeout should be enough to deal
with random network hiccups, and it is unwise to indiscriminately retry
remote command execution as we don't know if the command has already
been issued before the connection was severed.