Fix boot connection issues on LXD VMs and OCI
LXD VMs reboot mid-initialization, which can happen after we've already
connected to SSH. Much of the info reported by LXD isn't available until
after the VM is fully initialized, so rely of the reporting of number of
processes to determine if the we've fully booted.

OCI can have issues with writes being lost on shutdown. This is
particuarly bad on first boot and you lose the SSH keys in /etc/ssh.
Ensure we've issued a few sync's before we shutdown.