azure: perform smoke test to validate CLI config
Azure recent identity changes no longer raise an error
related to CLI misconfiguration/absence on initial client creation.
We need to minimally perform an operation against the CLI to confirm
presence or absence of necessary CLI config or fallback to using the
ClientSecretCredential.

Add smoke test to attempt to list operations of the resource client
if credentials are unavailable catch the ClientAuthenticationError
and fallback to ClientSecretCredential.