lxd: delete ephemeral instances using lxc delete
There are cases where ephemeral instances are not running and not
automatically deleted, e.g. when they are in ERROR state. Instances
in ERROR state can't be stopped, so the shutdown() call fails.

Let's just call `lxc delete` as we do for non-ephemeral images, and
trust lxd to do the right thing.

This is a partial revert of 0869d07b7894d985e183be957ae8318aae667c00.