use more specific exception (#234)
In efcdf67, daily_image() was changed to raise an
Exception when an image is not found. Before that
commit this code path was raising a ValueError
later in _streams_query().

This change broke cloud-init integration test
functionality on EC2 which allowed booting
arbitrary AMIs. Switch the raised exception back to
ValueError.