Improve oracle example script's ease of use and rename it so that it can actually be run (#349)
Previously, if you try to run the oracle example script, it would throw
an error because the file is named `oci.py` and this conflicts with the
naming of the `oci` module such that pycloudlib would fail to import the
`oci` module. Thus, it has been renamed to `oracle.py`.

Also, the flexibility and ease of use of the script has been improved so
that the `compartment_id` and `availability_domain` values can be set
either in variables in the script, or passed as cli args. And if neither
of these two things are done, then the script will simply prompt the
user for the two values.