feat(oracle): ability to specify which fault domain to launch in (#368)
Fault domains are physical groupings of hardware within an availability
domain for when a compute instance is going to be created. PR allows to
specify which fault domain to use which is useful for low capacity
instances. If no fault domain is provided (the default) then oci infra
will automatically try to choose the best one.