refactor(gce): migrate discovery api to compute_v1
When pycloudlib was created, the compute engine library for python (compute_v1)
did not exist. As CPC is looking to make more contributions to the GCE module
of pycloudlib as part of an internal tooling refactor, it makes sense to move
to the more modern library that has support for type hinting and a more modern
development interface. The new library wraps the same underlying REST API, so
there is no regression in features.