feat!: add networking type and networking config typings
The enum NetworkingType allows for indicating IPv4 single stack, IPv6
singe stack, IPv4 & IPv6 (dual-stack), and also has the default of AUTO.
The default of "auto" indicates that the networking type is unknown or
when used to configure something on the cloud, like launch and instance
or attaching a NIC, it indicates that the current functionality that is
networking type unaware should be used. This will allow for easily
providing backwards compatability.

Currently this is not added to the base cloud or instance classes
anywhere. Further design decisions must be made before this is
implemented across the board.

BREAKING CHANGE: ImageType enum moved from pycloudlib.cloud to
pycloudlib.types. You must update this import when updating from
pycloudlib version <11.0.0 to >=11.0.0.