feat(oracle): specify which subnet an instance must use (#475)
The default behaviour for assigning subnets is to iterate over the VCN
and take the first available subnet then assign it. The following
changes allow the ability to specifically set which subnet in the VCN
should be used for the instance. This is particularly helpful in a
Kubernetes environment where many subnets exist on a single VCN.