fix(oracle): update condition for setting public ips
The Oracle provider in pycloudlib will attempt to assign
a public IP to the instance. However, the current implementation
assumed that the subnet was always public but this may not be true.

To remedy this, we explicitly add checks that the subnet allows
public ips to be assigned, then we set that configuration in
the vnic details"